remove applications layout loader and add separate loaders for applications board and applications list pages
