+++ client/src/App.tsx
@@ -44,10 +44,10 @@ import {
 import { ForgotPasswordPage } from "./pages/ForgotPasswordPage";
 import { ResetPasswordPage } from "./pages/ResetPasswordPage";
 import ContactUsPage from "./pages/ContactUsPage";
-import { AppsLayout, appsLayoutLoader } from "./layouts/AppsLayout";
-import { AppsBoardPage } from "./pages/AppsBoardPage";
+import { AppsLayout } from "./layouts/AppsLayout";
+import { AppsBoardPage, appsBoardLoader } from "./pages/AppsBoardPage";
 import { BoardProvider } from "./context/BoardContext";
-import { AppsListPage } from "./pages/AppsListPage";
+import { AppsListPage, appsListLoader } from "./pages/AppsListPage";
 
 const router = createBrowserRouter([
   {
@@ -102,7 +102,6 @@ const router = createBrowserRouter([
       {
         path: "/:username/apps",
         element: <AppsLayout />,
-        loader: appsLayoutLoader,
         id: "apps",
         children: [
           {
@@ -116,10 +115,12 @@ const router = createBrowserRouter([
                 <AppsBoardPage />
               </BoardProvider>
             ),
+            loader: appsListLoader,
           },
           {
             path: "list",
             element: <AppsListPage />,
+            loader: appsBoardLoader,
           },
         ],
       },
