+++ client/src/context/BoardContext.tsx
@@ -1,12 +1,12 @@
 import React, { useState, createContext, useContext } from "react";
-import { useRouteLoaderData } from "react-router-dom";
+import { useLoaderData } from "react-router-dom";
 import { UserType } from "../../type-definitions";
 
 const BoardContext = createContext<any>({});
 
 export const useBoard = () => useContext(BoardContext);
 export const BoardProvider = ({ children }: { children: any }) => {
-  const { userData } = useRouteLoaderData("apps") as {
+  const { userData } = useLoaderData() as {
     userData: UserType;
   };
   const [boardData, setBoardData] = useState<any>({
