+++ client/src/layouts/AppsLayout.tsx
@@ -1,25 +1,16 @@
 import React from "react";
-import {
-  NavLink,
-  Navigate,
-  Outlet,
-  Params,
-  useParams,
-  useRouteLoaderData,
-} from "react-router-dom";
+import { NavLink, Navigate, Outlet, useParams } from "react-router-dom";
 import AuthedPageTitle from "../components/AuthedPageTitle";
-import axios from "axios";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faChalkboard, faList } from "@fortawesome/free-solid-svg-icons";
 import { useAuth } from "../context/AuthContext";
-import { UserType } from "../../type-definitions";
 
 export const AppsLayout = () => {
   const { authedUser } = useAuth();
   const { username } = useParams();
-  const { userData } = useRouteLoaderData("apps") as {
-    userData: UserType;
-  };
+
+  console.log(authedUser);
+
   const isAuthorizedUser = authedUser?.username === username;
 
   const activateSidebarLinks = ({ isActive }: { isActive: boolean }) => {
@@ -34,7 +25,7 @@ export const AppsLayout = () => {
     <>
       <AuthedPageTitle
         links={[
-          { to: `/${userData.username}`, label: userData.username },
+          { to: `/${authedUser?.username}`, label: authedUser?.username },
           { to: "", label: "MyApps" },
         ]}
       />
@@ -46,21 +37,22 @@ export const AppsLayout = () => {
                 className="w-7 h-7 rounded-full mr-3 sm:w-10 sm:h-10"
                 width={40}
                 height={40}
-                alt={userData.username}
+                alt={authedUser?.username}
                 src={authedUser?.photo || authedUser?.avatar}
               />
               <h1 className="text-base sm:text-2xl">
                 <NavLink
-                  to={`/${userData.username}`}
+                  to={`/${authedUser?.username}`}
                   className="no-underline font-semibold text-primary hover:underline"
                 >
-                  {userData.firstName}
-                  <span className="text-slate-600"> ({userData.username})</span>
+                  <span className="text-slate-600">
+                    ({authedUser?.username})
+                  </span>
                 </NavLink>
               </h1>
             </div>
             <NavLink
-              to={`/${userData.username}`}
+              to={`/${authedUser?.username}`}
               className="no-underline font-semibold text-sm min-w-fit text-primary p-2 bg-secondary rounded-md
             border border-slate-400 hover:border-slate-600 hover:bg-highlight sm:text-base"
             >
@@ -100,25 +92,4 @@ export const AppsLayout = () => {
       </div>
     </>
   );
-};
-
-export const appsLayoutLoader = async ({
-  request,
-  params,
-}: {
-  request: Request;
-  params: Params;
-}) => {
-  const url = new URL(request.url);
-  const searchParams = new URLSearchParams(url.search);
-  const appsListParams = {
-    sort: searchParams.get("sort"),
-    search: searchParams.get("search"),
-    page: searchParams.get("page"),
-  };
-  const userResponse = await axios.get("/api/users/user", {
-    params: appsListParams,
-  });
-  const userData = userResponse.data;
-  return { userData };
 };
