+++ client/src/pages/AppsBoardPage.tsx
@@ -22,6 +22,7 @@ import useOnClickOutside from "../hooks/useOnClickOutside";
 import { useBoard } from "../context/BoardContext";
 import BoardAppDetailsModal from "../components/BoardAppDetailsModal";
 import CreateBoardAppModal from "../components/CreateBoardAppModal";
+import { Params } from "react-router-dom";
 
 export const AppsBoardPage = () => {
   const {
@@ -332,4 +333,16 @@ export const AppsBoardPage = () => {
       </div>
     </>
   );
+};
+
+export const appsBoardLoader = async ({
+  request,
+  params,
+}: {
+  request: Request;
+  params: Params;
+}) => {
+  const userResponse = await axios.get("/api/users/user");
+  const userData = userResponse.data;
+  return { userData };
 };
